Initial commit, Desktop version finished no I
This commit contains the initial and the finished version of the desktop
website it also has links for the figma desing, canva presentation, etc.

However this commit doesn't contain the mobile responsive designs, or
the i icon in the bottom right showing that it was made by me, ward,
sohaila, nagham and hanna for DECI L3 summer bootcamp.

On branch master

Initial commit

Changes to be committed:
      new file:   css/main.css
      new file:   index.html
      new file:   js/main.js
      new file:   media/main.jpg
      new file:   media/s1.jpg
      new file:   media/s2.jpg
      new file:   media/s3.jpg
      new file:   media/s4.webp